+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Position of erroneously found references were wrong
+- `Find Usage`, `Resolve`, and  `Documentation` did not consider injected language
 
 ## [0.2.0] 2023-07-29